Fix complexity issues.
