+++ bin/app.php
@@ -1,32 +1,29 @@
 #!/usr/bin/env php
 <?php
 
+use Consolidation\AnnotatedCommand\CommandFileDiscovery;
+use Symfony\Component\Console\Output\ConsoleOutput;
+
 require_once __DIR__ . '/../vendor/autoload.php';
 
 $dktl_directory = DkanTools\Util\Util::getDktlDirectory();
 $dktl_project_directory = DkanTools\Util\Util::getProjectDirectory();
 
-$output = new \Symfony\Component\Console\Output\ConsoleOutput();
+$output = new ConsoleOutput();
 
-$discovery = new \Consolidation\AnnotatedCommand\CommandFileDiscovery();
+$discovery = new CommandFileDiscovery();
 $discovery->setSearchPattern('*Commands.php');
 $defaultCommandClasses = $discovery->discover("{$dktl_directory}/src", '\\DkanTools');
-
-$discovery->setSearchPattern('*Commands.php');
-$commandsClasses = $discovery->discover(
-    "{$dktl_directory}/src/Drupal",
-    '\\DkanTools\\Drupal\\'
-);
 
 $customCommandClasses = [];
 if (file_exists("{$dktl_project_directory}/src/command")) {
     $customCommandClasses = $discovery->discover("{$dktl_project_directory}/src/command", '\\DkanTools\\Custom');
 }
 
-$commandClasses = array_merge($defaultCommandClasses, $commandsClasses, $customCommandClasses);
+$commandClasses = array_merge($defaultCommandClasses, $customCommandClasses);
 
 $appName = "DKAN Tools";
-$appVersion = '1.0.0-alpha1';
+$appVersion = '2.0.0-rc1';
 $configurationFilename = 'dktl.yml';
 
 $runner = new \Robo\Runner($commandClasses);
@@ -34,7 +31,7 @@ $runner->setConfigurationFilename($configurationFilename);
 
 $argv = $_SERVER['argv'];
 
-$output = new \Symfony\Component\Console\Output\ConsoleOutput();
+$output = new ConsoleOutput();
 $statusCode = $runner->execute($argv, $appName, $appVersion, $output);
 
 exit($statusCode);
